41: Move under construction page as an ignore page route and move new home page to main / url. Resolves #41. WE ARE LIVE
